Restructure tests
Tests are now seperated according to type and the class that
being tested.
