+++ shade_janitor/tests/unit/resources/test_blacklist.py
@@ -1,15 +1,10 @@
-
 from mock import Mock
-from shade_janitor.resources import NoCloudException
-from shade_janitor.resources import Resources
 from unittest import TestCase
 
+from shade_janitor.resources import Resources
 
-class TestResources(TestCase):
 
-    def test_resources_fails_no_cloud(self):
-        with self.assertRaises(NoCloudException):
-            Resources(None)
+class TestBlackList(TestCase):
 
     def test_blacklist_normal(self):
         r = Resources(Mock())
@@ -40,15 +35,3 @@ class TestResources(TestCase):
         instance = Mock()
         instance.name = 'rpm-mirror-01'
         self.assertTrue(r.is_blacklisted(instance))
-
-    def test_is_permanent_non_perm_instance(self):
-        r = Resources(Mock())
-        instance = Mock()
-        instance.name = 'abcde-rdo-ci-88-foo'
-        self.assertFalse(r.is_permanent(instance))
-
-    def test_is_permanent_perm_instance(self):
-        r = Resources(Mock())
-        instance = Mock()
-        instance.name = 'permanent-rdo-ci-88-foo'
-        self.assertTrue(r.is_permanent(instance))
